requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
idna==3.3
    # via requests
joblib==1.1.0
    # via scikit-learn
lxml==4.7.1
    # via -r requirements.in
numpy==1.21.5
    # via
    #   pandas
    #   scikit-learn
    #   scipy
oauthlib==3.1.1
    # via requests-oauthlib
pandas==1.3.5
    # via -r requirements.in
psycopg2-binary==2.9.2
    # via -r requirements.in
pycodestyle==2.8.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2021.3
    # via pandas
requests==2.26.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via -r requirements.in
scikit-learn==1.0.2
    # via sklearn
scipy==1.7.3
    # via scikit-learn
six==1.16.0
    # via python-dateutil
sklearn==0.0
    # via -r requirements.in
threadpoolctl==3.0.0
    # via scikit-learn
urllib3==1.26.7
    # via requests
